Split localAuth into focused route modules (444 → 46 lines)
The single setupLocalAuth function was registering 9 routes + the
session middleware + the guards in one place. Each concern now has
its own file under server/local-auth/:

- types.ts                     LocalUser
- session-helpers.ts           VERIFICATION_ATTEMPTS rate-limit map,
                               getClientIp, clearAttemptsForIp,
                               sessionUser, attachUser
- session-middleware.ts        getSessionMiddleware — express-session
                               config from admin settings (cookie
                               hardening for hosted cross-origin
                               deploys)
- middleware.ts                ensureAuthenticatedSession guard +
                               isAuthenticated / isAdmin /
                               isLocalAuthenticated exports
- routes-login.ts              POST /api/login (username/password
                               OR admin secure-phrase) + /api/logout
- routes-admin-otp.ts          admin email OTP request/verify + the
                               public /api/admin/login/email hint
- routes-credentials.ts        /api/admin/verify-challenge (lockout
                               recovery) + self-service
                               /api/auth/update-credentials and
                               /api/auth/current-credentials
- routes-security-settings.ts  /api/admin/security-settings GET + POST

localAuth.ts is now a 46-line shell that re-exports the middleware
so all 10+ external import sites (route modules, the main routes
entry, etc.) keep working without changes.
