+++ server/local-auth/types.ts
@@ -1,0 +1,10 @@
+export interface LocalUser {
+  id: string;
+  username: string;
+  email: string;
+  firstName: string;
+  lastName: string;
+  profileImageUrl: string;
+  isAdmin: boolean;
+  isActive: boolean;
+}
